+++ src/py_pastream.c
@@ -5,7 +5,6 @@
 
 void init_stream(
     Py_PaBufferedStream *stream, 
-    PaStreamCallbackFlags allow_xruns, 
     unsigned char allow_drops,
     long frames,
     long pad,
@@ -13,7 +12,6 @@ void init_stream(
     PaUtilRingBuffer *rxbuff,
     PaUtilRingBuffer *txbuff) 
 {
-    stream->allow_xruns = allow_xruns;
     stream->allow_drops = allow_drops;
     stream->frames = (frames >= 0 && pad >= 0) ? (frames + pad) : frames;
     // technically if frames < 0 padding has no meaning, but set it anyways to
@@ -66,10 +64,6 @@ int callback(
             stream->outputUnderflows++;
         if ( status & paOutputOverflow )
             stream->outputOverflows++;
-        if ( status & ~(stream->allow_xruns | paPrimingOutput) ) {
-            strcpy(stream->errorMsg, "XRunError");
-            return (stream->last_callback = paAbort);
-        }
     }
 
     // exit point (1 of 2)
@@ -110,7 +104,7 @@ int callback(
                 // else { pad indefinitely; }
             }
             else if ( !stream->__autoframes && pad >= 0 ) {
-                strcpy(stream->errorMsg, "TransmitBufferEmpty");
+                strcpy(stream->errorMsg, "BufferEmpty");
                 stream->frame_count += oframes;
                 return (stream->last_callback = paAbort);
             }
@@ -126,7 +120,7 @@ int callback(
 
         iframes = PaUtil_WriteRingBuffer(stream->rxbuff, (const void *) in_data, frames_left);
         if ( iframes < frames_left && !stream->allow_drops ) {
-            strcpy(stream->errorMsg, "ReceiveBufferFull");
+            strcpy(stream->errorMsg, "BufferFull");
             stream->frame_count += iframes;
             return (stream->last_callback = paAbort);
         }